This is a bit of an experimental change.  Added coordinate sorting for blocks in chunks and disabled depth buffering for translucent blocks. This adds some extra computation, but makes translucent blocks such as stained glass look a lot better, especially with other translucent blocks behind it.
Fixes the inconsistency when looking at translucent blocks from
different angles.

Still have a problem with fluids within translucent blocks and some edge
cases, may be due to this not being an actual vertex sorter, but on the
block coordinate.

This sorting is also done for opaque objects as well unfortunately.
Will dig into optimizing in the future.