Finally fix the issue with WorldSlice, went back to using a blockStates array with index using regular multiplication instead of bitshifts and array length a power of 2.
This reverts a part of the optimization from
https://github.com/CaffeineMC/sodium-fabric/commit/eb664ec9bf22428678691f761fa0a6a73c916410
but prevents the NPE that was being hit.

My guess is the NPE was due to a mod compatibility issue when we stopped
using the regular ChunkSection.getBlockState() and tried fetching object
references from the fields.

This change may be very slightly less optimal, but it will be more compatible.